Make Items array a property of the view controller
That is tidier than leaving it as a global variable outside of the
class.

In large programs it is generally a good idea to avoid global variables
for a few reasons: 1) they can be accessed from anywhere, making it more
difficult to understand a program, 2) they discourage modular
programming by allowing code to rely on other code in a haphazard
fashion, making future maintenance of a program much more difficult, and
3) in a multi-threaded program, they must be handled very carefully or
they risk being accessed simultaneously by more than one thread
thereby introducing difficult-to-find non-deterministic bugs.